% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-RiemannSPD.R
\docType{package}
\name{package-RiemannSPD}
\alias{package-RiemannSPD}
\title{RiemannSPD}
\description{
We provide a variety of algorithms for symmetric and positive-definite (SPD) matrices, 
a collection of which has rich geometric characteristics.
}
