\name{RiemannSPD-package}
\alias{RiemannSPD-package}
\alias{RiemannSPD}
\docType{package}
\title{\packageTitle{RiemannSPD}}
\description{\packageDescription{RiemannSPD}}
\section{Package Content}{\packageIndices{RiemannSPD}}
\author{\packageAuthor{RiemannSPD}}
\section{Maintainer}{\packageMaintainer{RiemannSPD}}
\keyword{package}
