% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDR_cmds.R
\name{spd.cmds}
\alias{spd.cmds}
\title{Classical Multidimensional Scaling}
\usage{
spd.cmds(spd, geometry, ndim = 2)
}
\arguments{
\item{spd}{a S3 \code{"spd"} class for \eqn{N} of \eqn{(p\times p)} SPD matrices.}

\item{geometry}{(case-insensitive) name of supported geometry from \code{spd.geometry("spd.cmds")}.}

\item{ndim}{an integer-valued target dimension (default: 2).}
}
\value{
a named list containing \describe{
\item{embed}{an \eqn{(N\times ndim)} matrix whose rows are embedded observations.}
\item{stress}{discrepancy between embedded and original distances as a measure of error.}
}
}
\description{
Given \eqn{N} observations \eqn{\Sigma_1, \ldots, \Sigma_N} in the SPD manifold, 
apply classical multidimensional scaling to get low-dimensional representation 
in Euclidean space. The computation is based on pairwise dissimilarity, which is 
controlled by the \code{geometry} parameter.
}
\references{
\insertRef{torgerson_multidimensional_1952a}{RiemannSPD}
}
\concept{LDR}
