% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_gen2class.R
\name{spd.gen2class}
\alias{spd.gen2class}
\title{Generate 2 Types of SPD Matrices}
\usage{
spd.gen2class(p = 5, n1 = 10, n2 = 10, corr = FALSE)
}
\arguments{
\item{p}{dimensionality for SPD matrices (default: 5).}

\item{n1}{the number of type 1 SPD matrices (default: 10).}

\item{n2}{the number of type 2 SPD matrices (default: 10).}

\item{corr}{a logical; \code{TRUE} to coerce unit diagonal structure and \code{FALSE} otherwise (default: \code{FALSE}).}
}
\value{
a named list containing\describe{
\item{spd}{a S3 \code{spd} class object of \eqn{(n_1+n_2)} SPD matrices.}
\item{lab}{a length \eqn{(n_1 + n_2)} vector of class labels of numeric values.}
}
}
\description{
This function simulates two types of covariance matrices as follows. Given a model 
covariance matrix \eqn{\Sigma \in \mathbf{R}^{p\times p}}, a random sample of size 
\eqn{3p} is generated from multivariate normal model \eqn{\mathcal{N}(0,\Sigma)} and 
its empirical covariance \eqn{\hat{\Sigma}} is computed as a perturbed version of the model 
matrix. In this function, the first model matrix is an identity matrix \eqn{\Sigma_1 = I_{p}}. 
The second model matrix is generated from an empirical covariance of a matrix filled with 
random numbers drawn from a uniform distribution \eqn{\mathcal{U}(0,1)}.
}
\examples{
\donttest{
## simple example 
gen2obj = spd.gen2class(n1=5, n2=5)
gen_spd = gen2obj$spd     # 'spd' class object
gen_lab = gen2obj$lab     # corresponding label

## visualize all matrices
opar <- par(no.readonly=TRUE)
par(mfrow=c(2,5), pty="s")
for (i in 1:10){
  image(gen_spd$data[,,i], xaxt="n", yaxt="n",
        main=paste0("class ",gen_lab[i]))
}
par(opar)
}

}
\concept{utility}
