% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_geometry.R
\name{spd.geometry}
\alias{spd.geometry}
\title{Show available geometries for a chosen function}
\usage{
spd.geometry(fname)
}
\description{
Show available geometries for a chosen function
}
\concept{utility}
