% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmony_14N.R
\name{spd.harmony14N}
\alias{spd.harmony14N}
\title{Data Harmonization by Ng et al. (2014)}
\usage{
spd.harmony14N(spdlist, ...)
}
\arguments{
\item{spdlist}{a length-\eqn{K} list whose elements are S3 \code{"spd"} class objects.}

\item{...}{extra parameters including \describe{
\item{maxiter}{maximum number of iterations to be run (default:100).}
\item{eps}{tolerance level for stopping criterion (default: 1e-8).}
}}
}
\value{
a named list containing \describe{
\item{transformed}{a length-\eqn{K} list of harmonized dataset wrapped as \code{"spd"} class.}
\item{mean.local}{a length-\eqn{K} list of per-class/site mean.}
}
}
\description{
Data Harmonization by Ng et al. (2014)
}
\references{
\insertRef{ng_2014_TransportRiemannianManifold}{RiemannSPD}
}
\concept{harmony}
