% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmony_19Y.R
\name{spd.harmony19Y}
\alias{spd.harmony19Y}
\title{Data Harmonization by Yair et al. (2019)}
\usage{
spd.harmony19Y(spdlist, ...)
}
\description{
Data Harmonization by Yair et al. (2019)
}
\references{
\insertRef{yair_2019_harmony}{RiemannSPD}
}
\concept{harmony}
