% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmony_22S.R
\name{spd.harmony22S}
\alias{spd.harmony22S}
\title{Data Harmonization by Simeon et al. (2022)}
\usage{
spd.harmony22S(spdlist, ref = c("mean", "identity"))
}
\description{
Data Harmonization by Simeon et al. (2022)
}
\references{
\insertRef{simeon_2022_RiemannianGeometryFunctional}{RiemannSPD}
}
\concept{harmony}
