% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_kmedoids.R
\name{spd.kmedoids}
\alias{spd.kmedoids}
\title{K-medoids clustering}
\usage{
spd.kmedoids(spd, geometry, k = 2)
}
\arguments{
\item{spd}{a S3 \code{"spd"} class for \eqn{N} SPD matrices.}

\item{geometry}{(case-insensitive) name of supported geometry from \code{spd.geometry("spd.kmedoids")}.}

\item{k}{the number of clusters (default: 2). If \eqn{k<2}, it returns an error.}
}
\value{
a named list containing\describe{
\item{medoids}{a length-\eqn{k} vector of medoids' indices.}
\item{cluster}{a length-\eqn{N} vector of class labels (from \eqn{1:k}).}
}
}
\description{
Given \eqn{N} observations \eqn{\Sigma_1, \ldots, \Sigma_N} in the SPD manifold, 
perform \eqn{k}-medoids clustering algorithm. The computation is based on pairwise dissimilarity, which is 
controlled by the \code{geometry} parameter.
}
\concept{cluster}
