% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_pdist.R, R/basic_pdist2.R
\name{spd.pdist}
\alias{spd.pdist}
\title{Pairwise distance}
\usage{
spd.pdist(spd1, spd2, geometry)

spd.pdist(spd1, spd2, geometry)
}
\arguments{
\item{spd1}{a S3 \code{"spd"} class for \eqn{M} SPD matrices.}

\item{spd2}{a S3 \code{"spd"} class for \eqn{N} SPD matrices.}

\item{geometry}{(case-insensitive) name of supported geometry from \code{spd.geometry("spd.pdist2")}.}

\item{spd}{a S3 \code{"spd"} class for  \eqn{N} SPD matrices.}

\item{as.dist}{logical; if \code{TRUE}, it returns \code{dist} object, else it returns a symmetric matrix. (default: \code{FALSE}).}
}
\value{
a S3 \code{dist} object or \eqn{(N\times N)} symmetric matrix of pairwise distances according to \code{as.dist} parameter.

an \eqn{(M\times N)} matrix of distances.
}
\description{
Pairwise distance

Pairwise distance for two sets of data
}
\concept{basic}
