% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_pdist2.R
\name{spd.pdist2}
\alias{spd.pdist2}
\title{Pairwise distance for two sets of data}
\usage{
spd.pdist2(spd1, spd2, geometry)
}
\arguments{
\item{spd1}{a S3 \code{"spd"} class for \eqn{M} of \eqn{(p\times p)} SPD matrices.}

\item{spd2}{a S3 \code{"spd"} class for \eqn{N} of \eqn{(p\times p)} SPD matrices.}

\item{geometry}{(case-insensitive) name of supported geometry from \code{spd.geometry("spd.pdist2")}.}
}
\value{
an \eqn{(M\times N)} matrix of distances.
}
\description{
Given two sets of SPD matrices \eqn{X_1, \ldots, X_M} and \eqn{Y_1, \ldots, Y_N}, 
compute all pairwise distances across two collections of observations under 
the designated geometry.
}
\examples{
\donttest{
## toy data
spd_obj1 = spd.gen2class(n1=3, n2=2)$spd
spd_obj2 = spd.gen2class(n1=4, n2=6)$spd

## compute distance under different geometries
distA = spd.pdist2(spd_obj1, spd_obj2, "airm")
distL = spd.pdist2(spd_obj1, spd_obj2, "lerm")
distC = spd.pdist2(spd_obj1, spd_obj2, "chol")
distW = spd.pdist2(spd_obj1, spd_obj2, "wass")
 
## visualize different distance matrices
opar <- par(no.readonly=TRUE)
par(mfrow=c(2,2), pty="s")
image(distA, main="AIRM", xaxt="n", yaxt="n")
image(distL, main="LERM", xaxt="n", yaxt="n")
image(distC, main="Cholesky", xaxt="n", yaxt="n")
image(distW, main="Wasserstein", xaxt="n", yaxt="n")
par(opar)
}

}
\concept{basic}
