% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDR_pga.R
\name{spd.pga}
\alias{spd.pga}
\title{Principal Geodesic Analysis}
\usage{
spd.pga(spd, ndim = 2, ...)
}
\arguments{
\item{spd}{a S3 \code{"spd"} class for \eqn{N} of \eqn{(p\times p)} SPD matrices}

\item{ndim}{an integer-valued target dimension (default: 2).}

\item{...}{extra parameters including \describe{
\item{maxiter}{maximum number of iterations to be run (default:100).}
\item{eps}{tolerance level for stopping criterion (default: 1e-8).}
}}
}
\description{
Principal Geodesic Analysis
}
\concept{LDR}
