% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_wrap.R
\name{spd.wrap}
\alias{spd.wrap}
\title{Prepare Data on the SPD Manifold}
\usage{
spd.wrap(input, corr = FALSE)
}
\arguments{
\item{input}{SPD data matrices to be wrapped as \code{spd} class. Following inputs are considered,
\describe{
\item{array}{a \eqn{(p\times p\times n)} array where each slice along 3rd dimension is a SPD matrix.}
\item{list}{a length-\eqn{n} list whose elements are \eqn{(p\times p)} SPD matrices.}
}}

\item{corr}{a logical; \code{TRUE} to coerce unit diagonal structure and \code{FALSE} otherwise (default: \code{FALSE}).}
}
\value{
a named \code{spd} S3 object containing
\describe{
  \item{data}{a \eqn{(p\times p\times n)} 3d array of SPD matrices.}
  \item{size}{size of each SPD matrix.}
  \item{corr}{a logical for correlation-valued objects.}
}
}
\description{
The collection of symmetric positive-definite matrices is a well-known example 
of matrix manifold. It is defined as
\deqn{\mathcal{S}_{++}^p = \lbrace X \in \mathbf{R}^{p\times p} ~\vert~ X^\top = X,~ \textrm{rank}(X)=p \rbrace}
where the rank condition means it is strictly positive definite. In the \pkg{RiemannSPD}, 
we have a very strict policy to work with PD matrices only.
}
\examples{
\donttest{
# LOAD THE DATA
data(ERP)

# WRAP THE DATA
spdobj <- spd.wrap(ERP$spd)
}

}
\concept{prepare}
