% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_pdist.R
\name{spd.pdist}
\alias{spd.pdist}
\title{Pairwise distance}
\usage{
spd.pdist(spd, geometry, as.dist = FALSE)
}
\arguments{
\item{spd}{a S3 \code{"spd"} class for \eqn{N} of \eqn{(p\times p)} SPD matrices.}

\item{geometry}{(case-insensitive) name of supported geometry from \code{spd.geometry("spd.pdist")}.}

\item{as.dist}{logical; if \code{TRUE}, it returns \code{dist} object, else it returns a symmetric matrix. (default: \code{FALSE}).}
}
\value{
a S3 \code{dist} object or \eqn{(N\times N)} symmetric matrix of pairwise distances according to \code{as.dist} parameter.
}
\description{
Given a collection of SPD matrices \eqn{X_1, \ldots, X_N}, compute all 
pairwise distances under the designated geometry. Please see the following carefully;\describe{
\item{Note 1. \code{geometry="logdet"}}{
Since the square root of the Jensen-Bregman Log-Determinant divergence is 
metric, we return the square root values instead of the original divergences 
to be consistent with other \emph{metrics}.
}
}
}
\examples{
## toy data
gen2obj = spd.gen2class(n1=5, n2=5)
spd_obj = gen2obj$spd     # 'spd' class object

## compute distance under different geometries
distA = RiemannSPD::spd.pdist(spd_obj, "airm")
distL = spd.pdist(spd_obj, "lerm")
distC = spd.pdist(spd_obj, "chol")
distW = spd.pdist(spd_obj, "wass")
 
## visualize different distance matrices
opar <- par(no.readonly=TRUE)
par(mfrow=c(2,2), pty="s")
image(distA, main="AIRM", xaxt="n", yaxt="n")
image(distL, main="LERM", xaxt="n", yaxt="n")
image(distC, main="Cholesky", xaxt="n", yaxt="n")
image(distW, main="Wasserstein", xaxt="n", yaxt="n")
par(opar)

}
\concept{basic}
